pyproject.toml
[project]
name = "local-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "ipython>=9.6.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.34",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.8",
    "langsmith>=0.4.31",
    "mcp>=1.16.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "torch>=2.8.0",
]
